+++ jaseci_kit/jaseci_kit/cl_summer.py
@@ -1,1 +1,1 @@
-from .modules.summarization.summarization import *  # noqa
+from .modules.cl_summer.cl_summer import *  # noqa
