+++ jaseci_kit/jaseci_kit/ent_ext.py
@@ -1,1 +1,1 @@
-from .modules.entity_extraction.entity_extraction import *  # noqa
+from .modules.ent_ext.ent_ext import *  # noqa
