+++ jaseci_kit/jaseci_kit/modules/encoders/bi_enc.py
