+++ jaseci_kit/jaseci_kit/modules/encoders/tests/test_biencoder.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 from jaseci.utils.utils import TestCaseHelper
-from ..bi import serv_actions, config_setup
+from ..bi_enc import serv_actions, config_setup
 from fastapi.testclient import TestClient
 
 from .test_data import (
