+++ jaseci_kit/jaseci_kit/modules/ent_ext/tests/test_entity.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 from jaseci.utils.utils import TestCaseHelper
-from ..entity_extraction import serv_actions
+from ..ent_ext import serv_actions
 from fastapi.testclient import TestClient
 from .test_data import (
     test_entity_detection_request,
