+++ jaseci_kit/jaseci_kit/modules/fast_enc/fast_enc.py
