+++ jaseci_kit/jaseci_kit/t5_sum.py
@@ -1,1 +1,1 @@
-from .modules.t5_sum.t5_sum import *
+from .modules.t5_sum.t5_sum import *  # noqa
